+++ src/parameters/createOrganization.ts
@@ -7,8 +7,8 @@ export const CreateOrganizationSchema = z.object({
   desc: z.string().optional(),
   /**
    * A string with a length of at least 3. Only lowercase letters, underscores, and numbers are allowed. If the name
-   * contains invalid characters, they will be removed. If the name conflicts with an existing name, a new name will be
-   * substituted.
+   * contains invalid characters, they will be removed. If the name conflicts with an existing name, a new name will
+   * be substituted.
    */
   name: z.string().optional(),
   /** A URL starting with `http://` or `https://` */