+++ src/parameters/getBoard.ts
@@ -27,8 +27,8 @@ export const GetBoardSchema = z.object({
   customFields: z.boolean().optional(),
   /**
    * The fields of the board to be included in the response. Valid values: all or a comma-separated list of: closed,
-   * dateLastActivity, dateLastView, desc, descData, idMemberCreator, idOrganization, invitations, invited, labelNames,
-   * memberships, name, pinned, powerUps, prefs, shortLink, shortUrl, starred, subscribed, url
+   * dateLastActivity, dateLastView, desc, descData, idMemberCreator, idOrganization, invitations, invited,
+   * labelNames, memberships, name, pinned, powerUps, prefs, shortLink, shortUrl, starred, subscribed, url
    */
   fields: z.union([z.string(), z.array(z.string())]).optional(),
   /**