+++ src/parameters/markAllNotificationsRead.ts
@@ -5,8 +5,8 @@ export const MarkAllNotificationsReadSchema = z.object({
   read: z.boolean().optional(),
   /**
    * A comma-seperated list of IDs. Allows specifying an array of notification IDs to change the read state for. This
-   * will become useful as we add grouping of notifications to the UI, with a single button to mark all notifications in
-   * the group as read/unread.
+   * will become useful as we add grouping of notifications to the UI, with a single button to mark all notifications
+   * in the group as read/unread.
    */
   ids: z.array(z.string()).optional(),
 });